fix(metrics): Show memory metrics for zero
